move FilterDivider in component folder and publicly export
